feat: extract HTML diffing as dodeca-html-diff plugin
Move HTML DOM diffing from the main dodeca binary into a separate plugin,
reducing build dependencies and enabling parallel CI compilation.

- Create dodeca-html-diff plugin with diff_html() function
- Remove html5ever and markup5ever_rcdom from main binary dependencies
- Add plugin to "web" group in CI for parallel builds
- Graceful fallback to full page reload if plugin unavailable
