+++ crates/dodeca/src/lib.rs
@@ -5,7 +5,6 @@
 pub mod db;
 pub mod error_pages;
 pub mod file_watcher;
-pub mod html_diff;
 pub mod link_checker;
 pub mod plugins;
 pub mod template;
